Share some code in core
