+++ build.sbt
@@ -17,7 +17,7 @@ val http4sVersion = "0.21.4"
 val netty         = "4.1.50.Final"
 val munit         = "0.7.9"
 
-lazy val server = project.settings(
+lazy val core = project.settings(
   libraryDependencies ++= List(
     "co.fs2" %% "fs2-reactive-streams" % "2.3.0",
     ("com.typesafe.netty" % "netty-reactive-streams-http" % "2.0.4")
@@ -26,6 +26,12 @@ lazy val server = project.settings(
     "io.netty" % "netty-codec-http" % netty,
     ("io.netty" % "netty-transport-native-epoll" % netty).classifier("linux-x86_64"),
     ("io.netty" % "netty-transport-native-kqueue" % netty).classifier("osx-x86_64"),
+    "org.http4s" %% "http4s-core" % http4sVersion
+  )
+)
+
+lazy val server = project.dependsOn(core).settings(
+  libraryDependencies ++= List(
     "org.http4s" %% "http4s-server" % http4sVersion,
     "org.http4s" %% "http4s-dsl" % http4sVersion % Test,
     "ch.qos.logback" % "logback-classic" % "1.2.3" % Test,
@@ -36,15 +42,8 @@ lazy val server = project.settings(
   )
 )
 
-lazy val client = project.settings(
+lazy val client = project.dependsOn(core).settings(
   libraryDependencies ++= List(
-    "co.fs2" %% "fs2-reactive-streams" % "2.3.0",
-    ("com.typesafe.netty" % "netty-reactive-streams-http" % "2.0.4")
-      .exclude("io.netty", "netty-codec-http")
-      .exclude("io.netty", "netty-handler"),
-    "io.netty" % "netty-codec-http" % netty,
-    ("io.netty" % "netty-transport-native-epoll" % netty).classifier("linux-x86_64"),
-    ("io.netty" % "netty-transport-native-kqueue" % netty).classifier("osx-x86_64"),
     "org.http4s" %% "http4s-client" % http4sVersion,
     "org.scalameta" %% "munit" % munit % Test,
     "ch.qos.logback" % "logback-classic" % "1.2.3" % Test
