+++ client/src/main/scala/org/http4s/client/netty/NettyClientBuilder.scala
@@ -20,6 +20,7 @@ import org.http4s.{Request, Response}
 import org.http4s.Uri.Scheme
 import org.http4s.client.netty.NettyClientBuilder.SSLContextOption
 import org.http4s.client.{Client, RequestKey}
+import org.http4s.netty.{NettyChannelOptions, NettyTransport}
 
 import scala.concurrent.ExecutionContext
 import scala.concurrent.duration._
@@ -34,6 +35,7 @@ class NettyClientBuilder[F[_]](
     maxChunkSize: Int,
     transport: NettyTransport,
     sslContext: SSLContextOption,
+    nettyChannelOptions: NettyChannelOptions,
     executionContext: ExecutionContext
 )(implicit F: ConcurrentEffect[F]) {
   private[this] val logger = org.log4s.getLogger
@@ -48,9 +50,8 @@ class NettyClientBuilder[F[_]](
       maxChunkSize: Int = maxChunkSize,
       transport: NettyTransport = transport,
       sslContext: SSLContextOption = sslContext,
+      nettyChannelOptions: NettyChannelOptions = nettyChannelOptions,
       executionContext: ExecutionContext = executionContext
-      //nettyChannelOptions: NettyClientBuilder.NettyChannelOptions = nettyChannelOptions,
-      //sslConfig: NettyClientBuilder.SslConfig[F] = sslConfig
   ): NettyClientBuilder[F] =
     new NettyClientBuilder[F](
       idleTimeout,
@@ -60,14 +61,13 @@ class NettyClientBuilder[F[_]](
       maxChunkSize,
       transport,
       sslContext,
+      nettyChannelOptions,
       executionContext
-
-      //nettyChannelOptions,
-      //sslConfig
     )
 
   def withNativeTransport: Self = copy(transport = NettyTransport.Native)
   def withNioTransport: Self = copy(transport = NettyTransport.Nio)
+  def withExecutionContext(ec: ExecutionContext): Self = copy(executionContext = ec)
   def withMaxInitialLength(size: Int): Self = copy(maxInitialLength = size)
   def withMaxHeaderSize(size: Int): Self = copy(maxHeaderSize = size)
   def withMaxChunkSize(size: Int): Self = copy(maxChunkSize = size)
@@ -81,6 +81,16 @@ class NettyClientBuilder[F[_]](
 
   def withDefaultSSLContext: Self =
     copy(sslContext = NettyClientBuilder.SSLContextOption.TryDefaultSSLContext)
+
+  def withNettyChannelOptions(opts: NettyChannelOptions): Self =
+    copy(nettyChannelOptions = opts)
+
+  /**
+    * Socket selector threads.
+    * @param nThreads number of selector threads. Use <code>0</code> for netty default
+    * @return an updated builder
+    */
+  def withEventLoopThreads(nThreads: Int): Self = copy(eventLoopThreads = nThreads)
 
   private def getEventLoop: EventLoopHolder[_ <: SocketChannel] =
     transport match {
@@ -100,6 +110,8 @@ class NettyClientBuilder[F[_]](
   def setup: Bootstrap = {
     val bootstrap = new Bootstrap()
     getEventLoop.configure(bootstrap)
+    nettyChannelOptions.foldLeft(bootstrap) { case (boot, (opt, value)) => boot.option(opt, value) }
+
     bootstrap.handler(new ChannelInitializer[SocketChannel] {
       override def initChannel(channel: SocketChannel): Unit = {
         logger.trace(s"Initializing $channel")
@@ -122,9 +134,7 @@ class NettyClientBuilder[F[_]](
             engine.setUseClientMode(true)
             pipeline.addFirst("ssl", new SslHandler(engine))
             ()
-          case (m, a) =>
-            logger.trace(s"hmm $m, $a")
-            ()
+          case (_, _) => ()
         }
       }
     })
@@ -197,47 +207,9 @@ object NettyClientBuilder {
       maxChunkSize = 8192,
       transport = NettyTransport.Native,
       sslContext = SSLContextOption.TryDefaultSSLContext,
+      nettyChannelOptions = NettyChannelOptions.empty,
       executionContext = ExecutionContext.global
     )
-
-  /** Ensure we construct our netty channel options in a typeful, immutable way, despite
-    * the underlying being disgusting
-    */
-  sealed abstract class NettyChannelOptions {
-
-    /** Prepend to the channel options **/
-    def prepend[O](channelOption: ChannelOption[O], value: O): NettyChannelOptions
-
-    /** Append to the channel options **/
-    def append[O](channelOption: ChannelOption[O], value: O): NettyChannelOptions
-
-    /** Remove a channel option, if present **/
-    def remove[O](channelOption: ChannelOption[O]): NettyChannelOptions
-
-    private[http4s] def foldLeft[O](initial: O)(f: (O, (ChannelOption[Any], Any)) => O): O
-  }
-
-  object NettyChannelOptions {
-    val empty = new NettyCOptions(Vector.empty)
-  }
-
-  private[http4s] final class NettyCOptions(
-      private[http4s] val underlying: Vector[(ChannelOption[Any], Any)])
-      extends NettyChannelOptions {
-
-    def prepend[O](channelOption: ChannelOption[O], value: O): NettyChannelOptions =
-      new NettyCOptions((channelOption.asInstanceOf[ChannelOption[Any]], value: Any) +: underlying)
-
-    def append[O](channelOption: ChannelOption[O], value: O): NettyChannelOptions =
-      new NettyCOptions(
-        underlying :+ ((channelOption.asInstanceOf[ChannelOption[Any]], value: Any)))
-
-    def remove[O](channelOption: ChannelOption[O]): NettyChannelOptions =
-      new NettyCOptions(underlying.filterNot(_._1 == channelOption))
-
-    private[http4s] def foldLeft[O](initial: O)(f: (O, (ChannelOption[Any], Any)) => O) =
-      underlying.foldLeft[O](initial)(f)
-  }
 
   private sealed trait SSLContextOption extends Product with Serializable
   private object SSLContextOption {
@@ -258,11 +230,4 @@ object NettyClientBuilder {
         case NonFatal(_) => None
       }
   }
-}
-
-sealed trait NettyTransport extends Product with Serializable
-
-object NettyTransport {
-  case object Nio extends NettyTransport
-  case object Native extends NettyTransport
 }
