+++ core/src/main/scala/org/http4s/netty/NettyTransport.scala
@@ -1,0 +1,8 @@
+package org.http4s.netty
+
+sealed trait NettyTransport extends Product with Serializable
+
+object NettyTransport {
+  case object Nio extends NettyTransport
+  case object Native extends NettyTransport
+}
