+++ server/src/main/scala/org/http4s/server/netty/NettyServerBuilder.scala
@@ -19,6 +19,7 @@ import io.netty.handler.ssl.SslHandler
 import io.netty.handler.timeout.IdleStateHandler
 import javax.net.ssl.{SSLContext, SSLEngine}
 import org.http4s.HttpApp
+import org.http4s.netty.{NettyChannelOptions, NettyTransport}
 import org.http4s.server.{Server, ServiceErrorHandler, defaults}
 
 import scala.collection.immutable
@@ -38,7 +39,7 @@ final class NettyServerBuilder[F[_]](
     transport: NettyTransport,
     banner: immutable.Seq[String],
     executionContext: ExecutionContext,
-    nettyChannelOptions: NettyServerBuilder.NettyChannelOptions,
+    nettyChannelOptions: NettyChannelOptions,
     sslConfig: NettyServerBuilder.SslConfig[F],
     websocketsEnabled: Boolean,
     wsMaxFrameLength: Int
@@ -58,7 +59,7 @@ final class NettyServerBuilder[F[_]](
       transport: NettyTransport = transport,
       banner: immutable.Seq[String] = banner,
       executionContext: ExecutionContext = executionContext,
-      nettyChannelOptions: NettyServerBuilder.NettyChannelOptions = nettyChannelOptions,
+      nettyChannelOptions: NettyChannelOptions = nettyChannelOptions,
       sslConfig: NettyServerBuilder.SslConfig[F] = sslConfig,
       websocketsEnabled: Boolean = websocketsEnabled,
       wsMaxFrameLength: Int = wsMaxFrameLength
@@ -113,7 +114,7 @@ final class NettyServerBuilder[F[_]](
   def withMaxInitialLineLength(size: Int): Self = copy(maxInitialLineLength = size)
   def withServiceErrorHandler(handler: ServiceErrorHandler[F]): Self =
     copy(serviceErrorHandler = handler)
-  def withNettyChannelOptions(opts: NettyServerBuilder.NettyChannelOptions): Self =
+  def withNettyChannelOptions(opts: NettyChannelOptions): Self =
     copy(nettyChannelOptions = opts)
   def withWebsockets: Self = copy(websocketsEnabled = true)
   def withoutWebsockets: Self = copy(websocketsEnabled = false)
@@ -256,45 +257,6 @@ object NettyServerBuilder {
       wsMaxFrameLength = DefaultWSMaxFrameLength
     )
 
-  /** Ensure we construct our netty channel options in a typeful, immutable way, despite
-    * the underlying being disgusting
-    */
-  sealed abstract class NettyChannelOptions {
-
-    /** Prepend to the channel options **/
-    def prepend[O](channelOption: ChannelOption[O], value: O): NettyChannelOptions
-
-    /** Append to the channel options **/
-    def append[O](channelOption: ChannelOption[O], value: O): NettyChannelOptions
-
-    /** Remove a channel option, if present **/
-    def remove[O](channelOption: ChannelOption[O]): NettyChannelOptions
-
-    private[http4s] def foldLeft[O](initial: O)(f: (O, (ChannelOption[Any], Any)) => O): O
-  }
-
-  object NettyChannelOptions {
-    val empty = new NettyCOptions(Vector.empty)
-  }
-
-  private[http4s] final class NettyCOptions(
-      private[http4s] val underlying: Vector[(ChannelOption[Any], Any)])
-      extends NettyChannelOptions {
-
-    def prepend[O](channelOption: ChannelOption[O], value: O): NettyChannelOptions =
-      new NettyCOptions((channelOption.asInstanceOf[ChannelOption[Any]], value: Any) +: underlying)
-
-    def append[O](channelOption: ChannelOption[O], value: O): NettyChannelOptions =
-      new NettyCOptions(
-        underlying :+ ((channelOption.asInstanceOf[ChannelOption[Any]], value: Any)))
-
-    def remove[O](channelOption: ChannelOption[O]): NettyChannelOptions =
-      new NettyCOptions(underlying.filterNot(_._1 == channelOption))
-
-    private[http4s] def foldLeft[O](initial: O)(f: (O, (ChannelOption[Any], Any)) => O) =
-      underlying.foldLeft[O](initial)(f)
-  }
-
   private sealed trait SslConfig[F[_]] {
     def makeContext: F[Option[SSLContext]]
     def configureEngine(sslEngine: SSLEngine): Unit
@@ -317,11 +279,4 @@ object NettyServerBuilder {
     }
     def isSecure = false
   }
-}
-
-sealed trait NettyTransport extends Product with Serializable
-
-object NettyTransport {
-  case object Nio extends NettyTransport
-  case object Native extends NettyTransport
 }
